[Tech] clean code for lint
